% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_pkgs.R
\name{get_dataset_pkgs}
\alias{get_dataset_pkgs}
\title{Get CRAN R packages with datasets}
\usage{
get_dataset_pkgs(browse = TRUE, repos = getOption("repos")["CRAN"])
}
\arguments{
\item{browse}{Logical. If \code{TRUE}, open the resulting data frame in a
browser.}

\item{repos}{A string pointing to a CRAN instance. If \code{NULL} or
\code{NA} then the master CRAN at \url{https://cran.r-project.org} is used.}
}
\value{
A data frame with two columns.
\describe{
\item{pkgname}{Character. Name of the package.}
\item{description}{Character. A quick description of the package.}
}
}
\description{
Gets a data frame of CRAN R Packages that mention "dataset" or "data set"
in the Title field of their DESCRIPTION file.
}
